Introduce MultipartBuilder
